Refactored Layout Component into multiple files
