+++ src/components/Layout/fusTheme.js
@@ -1,0 +1,24 @@
+import { createMuiTheme } from 'material-ui/styles'
+const theme = createMuiTheme({
+  palette: {
+    primary: {
+      '50': '#21412a',
+      '100': '#21412a',
+      '200': '#21412a',
+      '300': '#21412a',
+      '400': '#21412a',
+      '500': '#21412a',
+      '600': '#21412a',
+      '700': '#21412a',
+      '800': '#21412a',
+      '900': '#21412a',
+      A100: '#21412a',
+      A200: '#21412a',
+      A400: '#21412a',
+      A700: '#21412a',
+      contrastDefaultColor: 'light'
+    }
+  }
+})
+
+export default theme
