+++ src/components/Layout/styles.js
@@ -1,0 +1,95 @@
+const drawerWidth = 250
+
+const styles = theme => ({
+  logo: {
+    width: '234px',
+    margin: '8px auto'
+  },
+  flex: {
+    flex: 1
+  },
+  link: {
+    textDecoration: 'none'
+  },
+  root: {
+    width: '100%',
+    height: 430,
+    zIndex: 1
+  },
+  appFrame: {
+    position: 'relative',
+    display: 'flex',
+    width: '100%',
+    height: '100%'
+  },
+  appBar: {
+    backgroundColor: '#fff',
+    borderBottom: 'solid 3px #998643',
+    color: '#21412a',
+    position: 'fixed',
+    transition: theme.transitions.create(['margin', 'width'], {
+      easing: theme.transitions.easing.sharp,
+      duration: theme.transitions.duration.leavingScreen
+    })
+  },
+  appBarShift: {
+    marginLeft: drawerWidth,
+    width: `calc(100% - ${drawerWidth}px)`,
+    transition: theme.transitions.create(['margin', 'width'], {
+      easing: theme.transitions.easing.easeOut,
+      duration: theme.transitions.duration.enteringScreen
+    })
+  },
+  menuButton: {
+    marginLeft: 12,
+    marginRight: 20
+  },
+  hide: {
+    display: 'none'
+  },
+  drawerPaper: {
+    left: 0,
+    position: 'fixed',
+    height: '100%',
+    width: drawerWidth
+  },
+  drawerHeader: {
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'flex-end',
+    padding: '0 8px',
+    ...theme.mixins.toolbar
+  },
+  content: {
+    width: '100%',
+    marginLeft: 0,
+    flexGrow: 1,
+    backgroundColor: theme.palette.background.default,
+    padding: theme.spacing.unit * 3,
+    transition: theme.transitions.create('margin', {
+      easing: theme.transitions.easing.sharp,
+      duration: theme.transitions.duration.leavingScreen
+    }),
+    height: 'calc(100% - 56px)',
+    marginTop: 56,
+    [theme.breakpoints.up('sm')]: {
+      content: {
+        height: 'calc(100% - 64px)',
+        marginTop: 64
+      }
+    }
+  },
+  contentShift: {
+    marginLeft: drawerWidth,
+    transition: theme.transitions.create('margin', {
+      easing: theme.transitions.easing.easeOut,
+      duration: theme.transitions.duration.enteringScreen
+    })
+  },
+  image: {
+    height: '100%',
+    width: '100%'
+  }
+})
+
+export default styles
