+++ src/views/Category.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 import { graphql } from 'react-apollo'
 import { getCategories } from '../graphql/queries/categories'
 import Loader from '../components/Loader'
-import Layout from '../components/Layout'
+import Layout from '../components/Layout/index'
 import CategoryView from '../components/CategoryView'
 import { Helmet } from 'react-helmet'
 // import { Link } from 'react-router-dom'
